Description: More stuff during updates. Reviewed by: dcid Bug:

+++ src/init/update.sh
@@ -111,9 +111,11 @@ UpdateStopOSSEC()
    $DIRECTORY/bin/ossec-control stop 
 
    # We also need to remove all syscheck queue file (format changed)
-   rm -f $DIRECTORY/queue/syscheck/* > /dev/null 2>&1
-   rm -f $DIRECTORY/queue/syscheck/.* > /dev/null 2>&1
-   rm -f $DIRECTORY/queue/agent-info/* > /dev/null 2>&1
+   if [ ! "X$VERSION" = "X0.9-3" ]; then
+        rm -f $DIRECTORY/queue/syscheck/* > /dev/null 2>&1
+        rm -f $DIRECTORY/queue/syscheck/.* > /dev/null 2>&1
+        rm -f $DIRECTORY/queue/agent-info/* > /dev/null 2>&1
+   fi
 }
 
 
@@ -129,7 +131,7 @@ UpdateOSSECRules()
     # Backing up the old config
     cp -pr ${OSSEC_CONF_FILE} "${OSSEC_CONF_FILE}.$$.bak"
     
-    cat ${OSSEC_CONF_FILE}|grep -v "rules" |grep -v "<include>" > "${OSSEC_CONF_FILE}.$$.tmp"
+    cat ${OSSEC_CONF_FILE}|grep -v "<rules>" |grep -v "<include>" > "${OSSEC_CONF_FILE}.$$.tmp"
 
     cat "${OSSEC_CONF_FILE}.$$.tmp" > ${OSSEC_CONF_FILE}
     rm "${OSSEC_CONF_FILE}.$$.tmp"
